Create yarn monorepo and split store and react utilities
