+++ .gitignore
@@ -8,3 +8,8 @@ es
 types
 
 yarn-error.log
+
+# Ignore any yarn package other than temperjs and temperjs-store
+packages/*
+!packages/temperjs
+!packages/temperjs-store
