+++ package.json
@@ -1,105 +1,8 @@
 {
+  "private": true,
   "name": "temperjs",
-  "version": "0.5.2",
-  "description": "State management for React, made simple.",
-  "homepage": "https://temperjs.org",
-  "repository": "github:ktripaldi/temperjs",
-  "bugs": "https://github.com/ktripaldi/temperjs/issues",
-  "keywords": [
-    "temper",
-    "temperjs",
-    "state",
-    "trait",
-    "react",
-    "reactjs",
-    "recoil",
-    "redux",
-    "flux"
+  "workspaces": [
+    "packages/*"
   ],
-  "main": "lib/temper.js",
-  "unpkg": "dist/temper.js",
-  "module": "es/temper.js",
-  "types": "types/index.d.ts",
-  "files": [
-    "dist",
-    "lib",
-    "es",
-    "types"
-  ],
-  "scripts": {
-    "clean": "rimraf lib dist es coverage types",
-    "check-types": "tsc --noEmit",
-    "build": "rollup -c",
-    "build:watch": "yarn run build --watch",
-    "test": "jest --config jestconfig.json --silent",
-    "test:watch": "yarn run test --watchAll",
-    "test:cov": "yarn run test --coverage",
-    "test:staged": "cross-env CI=true yarn run test --findRelatedTests",
-    "format": "prettier --write \"src/**/*.ts\" \"**/*.md\"",
-    "lint": "eslint --ext ts src"
-  },
-  "peerDependencies": {
-    "react": "^16.13.1"
-  },
-  "dependencies": {
-    "lodash-es": "^4.17.15"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.9.0",
-    "@babel/plugin-proposal-object-rest-spread": "^7.9.5",
-    "@babel/preset-env": "^7.9.5",
-    "@babel/preset-typescript": "^7.9.0",
-    "@babel/register": "^7.9.0",
-    "@rollup/plugin-babel": "^5.0.0",
-    "@rollup/plugin-node-resolve": "^7.1.3",
-    "@rollup/plugin-replace": "^2.3.2",
-    "@testing-library/jest-dom": "^5.11.3",
-    "@testing-library/react-hooks": "^3.4.1",
-    "@types/jest": "^25.2.1",
-    "@types/lodash-es": "^4.17.3",
-    "@types/react": "^16.9.35",
-    "@typescript-eslint/eslint-plugin": "^2.29.0",
-    "@typescript-eslint/parser": "^2.29.0",
-    "babel-eslint": "^10.1.0",
-    "babel-jest": "^25.4.0",
-    "cross-env": "^7.0.2",
-    "eslint": "^6.8.0",
-    "eslint-import-resolver-typescript": "^2.0.0",
-    "eslint-plugin-import": "^2.20.2",
-    "eslint-plugin-react": "^7.19.0",
-    "eslint-plugin-react-hooks": "^3.0.0",
-    "husky": "^4.3.0",
-    "jest": "^26.4.0",
-    "lint-staged": "^10.3.0",
-    "prettier": "^2.0.5",
-    "react": "^16.13.1",
-    "react-test-renderer": "^16.13.1",
-    "rimraf": "^3.0.2",
-    "rollup": "^2.7.2",
-    "rollup-plugin-terser": "^5.3.0",
-    "rollup-plugin-typescript2": "^0.27.0",
-    "rxjs": "^6.6.3",
-    "ts-jest": "^26.1.1",
-    "typescript": "^3.8.3"
-  },
-  "lint-staged": {
-    "src/**/*.{ts,tsx}": [
-      "eslint --fix",
-      "prettier --write",
-      "yarn test:staged"
-    ]
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged"
-    }
-  },
-  "author": "Kenderson Tripaldi",
-  "maintainers": [
-    {
-      "name": "Kenderson Tripaldi",
-      "email": "temperjs@icloud.com"
-    }
-  ],
-  "license": "MIT"
+  "scripts": {}
 }
