+++ packages/temperjs-store/rollup.config.js
@@ -1,0 +1,167 @@
+import nodeResolve from '@rollup/plugin-node-resolve'
+import babel from '@rollup/plugin-babel'
+import replace from '@rollup/plugin-replace'
+import typescript from 'rollup-plugin-typescript2'
+import { terser } from 'rollup-plugin-terser'
+
+import pkg from './package.json'
+
+const extensions = ['.ts']
+const noDeclarationFiles = { compilerOptions: { declaration: false } }
+
+const makeExternalPredicate = externalArr => {
+  if (externalArr.length === 0) {
+    return () => false
+  }
+  const pattern = new RegExp(`^(${externalArr.join('|')})($|/)`)
+  return id => pattern.test(id)
+}
+
+const external = makeExternalPredicate([
+  ...Object.keys(pkg.dependencies || {}),
+  ...Object.keys(pkg.peerDependencies || {})
+])
+
+const output = {
+  indent: false,
+  globals: {
+    'lodash-es': 'lodashEs'
+  }
+}
+
+export default [
+  // CommonJS
+  {
+    input: 'src/index.ts',
+    output: {
+      file: 'lib/temper-store.js',
+      format: 'cjs',
+      ...output
+    },
+    external,
+    plugins: [
+      nodeResolve({
+        extensions
+      }),
+      typescript({ useTsconfigDeclarationDir: true }),
+      babel({
+        extensions,
+        babelHelpers: 'bundled'
+      })
+    ]
+  },
+
+  // ES
+  {
+    input: 'src/index.ts',
+    output: {
+      file: 'es/temper-store.js',
+      format: 'es',
+      ...output
+    },
+    external,
+    plugins: [
+      nodeResolve({
+        extensions
+      }),
+      typescript({ tsconfigOverride: noDeclarationFiles }),
+      babel({
+        extensions,
+        babelHelpers: 'bundled'
+      })
+    ]
+  },
+
+  // ES for Browsers
+  {
+    input: 'src/index.ts',
+    output: {
+      file: 'es/temper-store.mjs',
+      format: 'es',
+      ...output
+    },
+    external,
+    plugins: [
+      nodeResolve({
+        extensions
+      }),
+      replace({
+        'process.env.NODE_ENV': JSON.stringify('production')
+      }),
+      typescript({ tsconfigOverride: noDeclarationFiles }),
+      babel({
+        extensions,
+        exclude: 'node_modules/**',
+        babelHelpers: 'bundled'
+      }),
+      terser({
+        compress: {
+          pure_getters: true,
+          unsafe: true,
+          unsafe_comps: true,
+          warnings: false
+        }
+      })
+    ]
+  },
+
+  // UMD Development
+  {
+    input: 'src/index.ts',
+    output: {
+      file: 'dist/temper-store.js',
+      format: 'umd',
+      name: 'Temperjs',
+      ...output
+    },
+    external,
+    plugins: [
+      nodeResolve({
+        extensions
+      }),
+      typescript({ tsconfigOverride: noDeclarationFiles }),
+      babel({
+        extensions,
+        exclude: 'node_modules/**',
+        babelHelpers: 'bundled'
+      }),
+      replace({
+        'process.env.NODE_ENV': JSON.stringify('development')
+      })
+    ]
+  },
+
+  // UMD Production
+  {
+    input: 'src/index.ts',
+    output: {
+      file: 'dist/temper-store.min.js',
+      format: 'umd',
+      name: 'Temperjs',
+      ...output
+    },
+    external,
+    plugins: [
+      nodeResolve({
+        extensions
+      }),
+      typescript({ tsconfigOverride: noDeclarationFiles }),
+      babel({
+        extensions,
+        exclude: 'node_modules/**',
+        babelHelpers: 'bundled'
+      }),
+      replace({
+        'process.env.NODE_ENV': JSON.stringify('production')
+      }),
+      terser({
+        compress: {
+          pure_getters: true,
+          unsafe: true,
+          unsafe_comps: true,
+          warnings: false
+        }
+      })
+    ]
+  }
+]
