+++ packages/temperjs-store/src/index.ts
@@ -1,0 +1,9 @@
+export { default as storeActions } from './services/store'
+export {
+  Trait,
+  SetterHelpers,
+  SetterValue,
+  Subscription,
+  StorageService,
+  StoreOptions
+} from './config/types'
