+++ packages/temperjs/src/actions/setTrait.ts
@@ -1,0 +1,14 @@
+import { storeActions } from 'temperjs-store'
+import { SetterValue } from '..'
+
+/**
+ * `setTrait` creates or updates the specified Trait
+ * @template T
+ * @param {string} path - This is the Trait identifier
+ * @param {SetterValue<T>} traitValue - This is the Trait value or the callback to set the value
+ */
+function setTrait<T>(path: string, traitValue: SetterValue<T>): void {
+  storeActions.setTrait<T>(path, traitValue)
+}
+
+export default setTrait
