+++ packages/temperjs/src/index.ts
@@ -1,0 +1,23 @@
+export { default as withTemper } from './hocs/withTemper'
+export { default as getTrait } from './actions/getTrait'
+export { default as setTrait } from './actions/setTrait'
+export { default as useTrait } from './hooks/useTrait'
+export { default as useTraitValue } from './hooks/useTraitValue'
+export {
+  Trait,
+  SetterHelpers,
+  SetterValue,
+  Subscription,
+  StorageService,
+  StoreOptions
+} from 'temperjs-store'
+export {
+  Loadable,
+  LoadableState,
+  SubscribedTrait,
+  Setter,
+  WithSetter,
+  AsyncWithSetter,
+  LoadableWithSetter,
+  SubscriptionOptions
+} from './config/types'
