--- src/index.ts
@@ -1,21 +1,0 @@
-export { default as withTemper } from './hocs/withTemper'
-export { default as getTrait } from './actions/getTrait'
-export { default as setTrait } from './actions/setTrait'
-export { default as useTrait } from './hooks/useTrait'
-export { default as useTraitValue } from './hooks/useTraitValue'
-export {
-  Trait,
-  Loadable,
-  LoadableState,
-  SubscribedTrait,
-  SetterHelpers,
-  SetterValue,
-  Setter,
-  WithSetter,
-  AsyncWithSetter,
-  LoadableWithSetter,
-  SubscriptionOptions,
-  Subscription,
-  StorageService,
-  StoreOptions
-} from './config/types'
